decompose(turn_bridge): extract S11 helper zones
